requirements.txt
# Resume Tailor AI - Dependencies
# Python 3.10+

# PDF text extraction
pypdf>=4.0.0

# OpenAI API
openai>=1.0.0

# Web app
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Environment
python-dotenv>=1.0.0

# Optional: HTML to PDF (server-side) - run: playwright install chromium
playwright>=1.40.0
